feat: init ansible vault file Reading secrets from the onepassword collection was incredibly slow, as it tried to fetch the secret for every task execution - a simple playbook could take upwards of 30 min to run!!
To get around this, we now use an ansible-vault encrypted file, whose
password is read from the 1password cli tool via a shell script. This
way, we reach out to 1password just once and our playbook's execution
remains speedy.

Note: The only annoying thing about this is that in editors like zed,
the onepassword secret appears to be read every time an ansible file is
read by the lsp at editor startup. I think we can get rid of if by
tweaking some setting in 1password, but this isn't a big issue for me,
esp on neovim. Battle for another day.